backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
chromadb>=0.4.22

# ML & AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
open-clip-torch>=2.24.0
insightface>=0.7.3
onnxruntime>=1.16.0

# Image/Video Processing
pillow>=10.2.0
opencv-python>=4.9.0.80
imageio>=2.33.1
imageio-ffmpeg>=0.4.9
exifread>=3.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
aiofiles>=23.2.1
tqdm>=4.66.1
numpy>=1.26.0

# Reverse Geocoding
geopy>=2.4.1
